+++ src/q1_time.py
@@ -76,4 +76,3 @@ def q1_time(file_path: str) -> List[Tuple[datetime.date, str]]:
         return result
 
 #q1_time("dbfs:/FileStore/shared_uploads/taboada.jose@pucp.pe/farmers_protest_tweets_2021_2_4.json")
-
