+++ src/q2_time.py
@@ -11,7 +11,7 @@ from typing import List, Tuple
 import re
 import emoji
 from pyspark.sql.functions import col, to_date, lit, count, row_number, substring, explode, split, udf
-from pyspark.sql.types import BooleanType
+from pyspark.sql.types import BooleanType, StructType, StringType, StructField
 from pyspark.sql import Window
 
 
@@ -22,32 +22,22 @@ def es_emoji(caracter):
     return emoji.is_emoji(caracter)
 
 
-def q2_memory(file_path: str) -> List[Tuple[str, int]]:
+def q2_time(file_path: str) -> List[Tuple[str, int]]:
     #Inicializamos spark
     spark = create_spark_session()
+    #Usamos solo el esquema solo con los campos necesarios para que el dataframe solo infiera esos campos y el tiempo de ejecución disminuya
+    schema = StructType([
+            StructField("content", StringType(), True)
+        ])
     #Lectura del json
-    df_json = read_json(spark, file_path)
-    #Obtenemos el campo a analizar
-    df_universo = df_json.select("content")
+    df_universo = read_json(spark, file_path)
 
-
-    
-    print(es_emoji("\u2764\ufe0f"))
     #Registramos la función como UDF en spark
     es_emoji_udf = udf(es_emoji, BooleanType())
 
     df_emojis = df_universo.select(explode(split(col('content'), '')).alias('caracteres')).filter(es_emoji_udf(col('caracteres')))
 
     df_emojis_3 = df_emojis.groupby('caracteres').agg(count(lit(1)).alias("ctd_emoji")).orderBy("ctd_emoji", ascending=False)
-    #window_spec = Window.partitionBy("caracteres")
-    #df_emojis_3 = df_emojis_2.select(col('caracteres'), 
-    #                                 col('ctd_emoji'), 
-    #                                 row_number().over(orderBy(col("ctd_emoji").desc())).alias("top_n"))
-    #df_emojis_3.display()
-
-    #df_emojis_4 = df_emojis_3.where(col('top_n') <= 10)
-
-    #df_emojis_4.display()
 
     #Almacenamos el resultado en el formato pedido
     result = write_to_tuples(df_emojis_3, "caracteres", "ctd_emoji")
@@ -55,4 +45,4 @@ def q2_memory(file_path: str) -> List[Tuple[str, int]]:
 
     return result
     
-q2_memory("dbfs:/FileStore/shared_uploads/taboada.jose@pucp.pe/farmers_protest_tweets_2021_2_4.json")
+q2_time("dbfs:/FileStore/shared_uploads/taboada.jose@pucp.pe/farmers_protest_tweets_2021_2_4.json")
