+++ src/q3_memory.py
@@ -45,4 +45,3 @@ def q3_memory(file_path: str) -> List[Tuple[str, int]]:
     return result
 
 q3_memory("dbfs:/FileStore/shared_uploads/taboada.jose@pucp.pe/farmers_protest_tweets_2021_2_4.json")
-
