+++ src/q3_time.py
@@ -52,4 +52,3 @@ def q3_time(file_path: str) -> List[Tuple[str, int]]:
     return result
 
 q3_time("dbfs:/FileStore/shared_uploads/taboada.jose@pucp.pe/farmers_protest_tweets_2021_2_4.json")
-
